control-electoral/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt>=4.0.0
python-multipart
pandas
psycopg2-binary
openpyxl
